server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
google-genai==1.53.0
python-dotenv==1.0.0
pydantic==2.12.5
pydub==0.25.1
psycopg2-binary==2.9.9
